# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APIs For Thrift Store",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/clientes": {
            "get": {
                "tags": [
                    "Clients"
                ],
                "summary": "Get a listing of the clients",
                "operationId": "29593cb6d735e27309ce747b0c9fbc61",
                "parameters": [
                    {
                        "name": "nombre",
                        "in": "query",
                        "description": "Filter by client name",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "paginacion",
                        "in": "query",
                        "description": "Number of rows per page",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/clientes/store": {
            "post": {
                "tags": [
                    "Clients"
                ],
                "summary": "Create a new client",
                "description": "Store a newly created client in storage.",
                "operationId": "6dfde32065b738f1faa21ef11b3b8c56",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "adress",
                                    "phone",
                                    "email"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "servicios_id": {
                                        "type": "id",
                                        "example": "1"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "123 Main St"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "555-123-4567"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "john@example.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                }
            }
        },
        "/api/clientes/{id}/show": {
            "get": {
                "tags": [
                    "Clients"
                ],
                "summary": "Get a specific client by ID",
                "description": "Display the specified client.",
                "operationId": "dffdc41d3cbfa0f36ee1b009bdc635f7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the client",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "nombre": {
                                            "type": "string",
                                            "example": "John Doe"
                                        },
                                        "address": {
                                            "type": "string",
                                            "example": "123 Main St"
                                        },
                                        "phone": {
                                            "type": "string",
                                            "example": "555-123-4567"
                                        },
                                        "email": {
                                            "type": "string",
                                            "format": "email",
                                            "example": "john@example.com"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Client not found"
                    }
                }
            }
        },
        "/api/clientes/{cliente}/update": {
            "put": {
                "tags": [
                    "Clients"
                ],
                "summary": "Update a client",
                "description": "Update the specified client in storage.",
                "operationId": "2aaaf98b20404c9416185a28f931478f",
                "parameters": [
                    {
                        "name": "cliente",
                        "in": "path",
                        "description": "ID of the client to update",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre",
                                    "address",
                                    "phone",
                                    "email"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "123 Main St"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "555-123-4567"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "john@example.com"
                                    },
                                    "servicios_id": {
                                        "type": "id",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Client updated successfully"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "404": {
                        "description": "Client not found"
                    }
                }
            }
        },
        "/api/clientes/{cliente}": {
            "delete": {
                "tags": [
                    "Clients"
                ],
                "summary": "Delete a client",
                "description": "Remove the specified client from storage.",
                "operationId": "863608963dfdca39c76ccca2457a8910",
                "parameters": [
                    {
                        "name": "cliente",
                        "in": "path",
                        "description": "ID of the client to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Client deleted successfully"
                    },
                    "404": {
                        "description": "Client not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}